Refactoring and adding some parms to the request
