+++ surplus-auctioneer-webapp/Views/Home/Index.cshtml
@@ -25,6 +25,6 @@
     <div class="col-md-4">
         <h2>Search Auction Items</h2>
         <p>Set your search criteria and find active auctions.</p>
-        <p><a class="btn btn-default" href="@Url.Action("Search")">Go &raquo;</a></p>
+        <p><a class="btn btn-default" href="@Url.Action("Search", "Search")">Go &raquo;</a></p>
     </div>
 </div>
