+++ surplus-auctioneer-webapp/Views/Shared/_Layout.cshtml
@@ -61,7 +61,7 @@
             <ul class="nav navbar-nav">
                 <li>@Html.ActionLink("Home", "Index", "Home")</li>
                 <li>@Html.ActionLink("Auctions", "Auctions", "Home")</li>
-                <li>@Html.ActionLink("Search", "Search", "Home")</li>
+                <li>@Html.ActionLink("Search", "Search", "Search")</li>
             </ul>
             <!--Comment this line out so that login isn't available
             @Html.Partial("_LoginPartial")-->
